Fix #2 Keep using the same output array
back-references are increasing, and can cross block lines.

Previously, an empty array was used for the next block. But
back-references were still high numbers (like 70000). Clearly, indexing
with such numbers fails, and it failed silently.

Now we use the same `ByteArray`  for all blocks. Not ideal from a speed
perspective, but at least it's correct...